% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_metadata.R
\name{call_cdec}
\alias{call_cdec}
\title{function makes call to and returns appropriate table from CDEC service}
\usage{
call_cdec(station)
}
\arguments{
\item{station}{three letter code for station to query for}
}
\description{
function makes call to and returns appropriate table from CDEC service
}
