% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_data.R
\name{call_cdec_avail_service}
\alias{call_cdec_avail_service}
\title{Function forms url and fetches the raw table from cdec avail service}
\usage{
call_cdec_avail_service(station)
}
\arguments{
\item{station}{station to retrieve available data for}
}
\value{
a raw table turned to dataframe from cdec service
}
\description{
Function forms url and fetches the raw table from cdec avail service
}
