% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdec_archive.R
\name{cdec_archive}
\alias{cdec_archive}
\title{Function extends retrieve_station_data and automatically places file in a
desired location}
\usage{
cdec_archive(station_id, sensor_num, dur_code, start_date, end_date,
  base_url = "shef", filename, comply_to = "psql")
}
\arguments{
\item{station_id}{three letter CDEC station code}

\item{sensor_num}{sensor number for desired data}

\item{start_date}{date for start of query}

\item{end_date}{last date to query data for}

\item{base_url}{url to retrieve data from, default is "shef" (most robust)}

\item{filename}{the destination for the archived file}

\item{comply_to}{defined a set of rules the csv must comply to, default is PostgreSQL}
}
\value{
NULL, function used for side-effects
}
\description{
Function extends retrieve_station_data and automatically places file in a
desired location
}
