% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{cdec_datasets_service}
\alias{cdec_datasets_service}
\title{Function forms url and fetches the raw table from cdec datasets "service"}
\usage{
cdec_datasets_service(station)
}
\arguments{
\item{station}{station to retrieve available data for}
}
\value{
a raw table turned to dataframe from cdec service
}
\description{
Function forms url and fetches the raw table from cdec datasets "service"
}
