% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_metadata.R
\name{cdec_historical_data_meta}
\alias{cdec_historical_data_meta}
\title{Get all sensor data for a given station.}
\usage{
cdec_historical_data_meta(station_id)
}
\arguments{
\item{station_id}{the three letter/number station id}
}
\value{
a data.frame with rows representing all the sensors with available data
}
\description{
Get all sensor data for a given station.
}
