% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{cdec_query}
\alias{cdec_query}
\alias{cdec_query.character}
\alias{cdec_query.cdec_dataset}
\title{Query observation data}
\usage{
cdec_query(x, ...)

\method{cdec_query}{character}(station, sensor_num, dur_code,
  start_date = NULL, end_date = NULL, tzone = "America/Los_Angeles")

\method{cdec_query}{cdec_dataset}(.data, row_id, start = NULL,
  end = NULL)
}
\arguments{
\item{station}{three letter identification for CDEC location (example "KWK", "SAC", "CCR")}

\item{sensor_num}{sensor number for the measure of interest. (example "20", "01", "25")}

\item{dur_code}{duration code for measure interval, "E", "H", "D", which correspong to Event, Hourly and Daily.}

\item{start_date}{date to start the query on.}

\item{end_date}{an optional date to end query on, defaults to current date.}

\item{tzone}{a time zone used. By default this is America/Los_Angeles, this accounts
for daylight saving.}

\item{row_id}{the row id from cdec_dataset output to query for}

\item{start}{change the start data for query}

\item{end}{change the end data for query}
}
\value{
dataframe

dataframe
}
\description{
Function queries the CDEC site to obtain desired station data
based on station, sensor number, duration code and start/end date.
Use cdec_datasets() to view an updated list of all available data at a station.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Query CDEC using known station, sensor and duration.

\item \code{cdec_dataset}: Use output from cdec_datasets to complete a query to CDEC.
}}

\examples{
\dontrun{
kwk_hourly_flows <- CDECRetrieve::cdec_query("KWK", "20", "H", "2017-01-01")
ccr_hourly_temps <- CDECRetrieve::cdec_query("CCR", "25", "H", Sys.Date())
}
}
