% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating-tables.R
\name{cdec_rt_list}
\alias{cdec_rt_list}
\title{List Rating Tables}
\usage{
cdec_rt_list(station_id = NULL)
}
\arguments{
\item{station_id}{station for the location to get rating description for.}
}
\value{
data frame with available rating tables as rows and columns describing
the \code{station_name}, \code{station_id}, data that data was \code{last_revised}, and \code{table_type.}
}
\description{
Get a list of all rating tables available through CDEC
}
\examples{
# list all rating tables in CDEC, you can use filter to search
\dontrun{
cdec_rt_list()
}
}
