% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdec_sensors}
\alias{cdec_sensors}
\title{The set of CDEC Codes as reported by CDEC}
\source{
\url{http://cdec.com}
}
\description{
A dataset containing all of the cdec codes, along with
other useful iformation including the units of measure for each
}
\keyword{data}
