% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdec_stations}
\alias{cdec_stations}
\title{CDEC Stations}
\format{dataframe with 2,559 rows and 9 variables
\describe{
\item{station_code}{three letter code used to identify the CDEC station}
\item{station_name}{long name for the station}
\item{elevation}{elevation in feet at the station}
\item{latitude}{latitude at station}
\item{longitude}{longitude at station}
\item{nearby_city}{nearby city to the station}
\item{hydro_code}{code used for the hydrological area}
\item{hydro_area}{hydrological area name}
\item{county}{county name for station}
}}
\usage{
cdec_stations
}
\description{
Data frame with all CDEC stations defined in the system
}
\keyword{datasets}
