% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consts.R
\docType{data}
\name{cdec_urls}
\alias{cdec_urls}
\title{Constant urls for data retrieval}
\format{An object of class \code{list} of length 4.}
\usage{
cdec_urls
}
\description{
Constant urls for data retrieval
}
\keyword{datasets}
