% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{clean_datasets_resp}
\alias{clean_datasets_resp}
\title{Function parses a response from CDEC available data service}
\usage{
clean_datasets_resp(df)
}
\description{
Function parses a response from CDEC available data service
}
