% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_data.R
\name{failed_query}
\alias{failed_query}
\title{A catch all for failed queries, helper function to below}
\usage{
failed_query(resp)
}
\arguments{
\item{resp}{a response from httr}
}
\value{
logical whether or not the response resulted in datasets
}
\description{
A catch all for failed queries, helper function to below
}
