% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_metadata.R
\name{get_station_metadata}
\alias{get_station_metadata}
\title{Use three letter code to retieve associated attributes}
\usage{
get_station_metadata(stations)
}
\arguments{
\item{stations}{a vector of three letter station codes.}
}
\value{
a list of attributes associated with each station in the \code{stations} vector
}
\description{
Use three letter code to retieve associated attributes
}
