% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_metadata.R
\name{get_station_table}
\alias{get_station_table}
\title{Function calls metadata service at CDEC and returns metadata for the corresponding station}
\usage{
get_station_table(station)
}
\arguments{
\item{station}{a three letter station code for metadata query}
}
\description{
Function calls metadata service at CDEC and returns metadata for the corresponding station
}
