% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_station_data.R
\name{make_cdec_url}
\alias{make_cdec_url}
\title{Function builds CDEC Url to request data}
\usage{
make_cdec_url(station_id, sensor_num, dur_code, start_date,
  end_date = as.character(Sys.Date()))
}
\arguments{
\item{station_id}{three letter identification for CDEC location}

\item{sensor_num}{sensor number for the measure of interest}

\item{dur_code}{duration code for measure interval, "E", "H", "D"}

\item{start_date}{date to start the query on}

\item{end_date}{a non-inclusive date to end the query on}
}
\value{
string url
}
\description{
Function builds CDEC Url to request data
}
