% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{map.cdec_stations}
\alias{map.cdec_stations}
\title{Map Station Search}
\usage{
\method{map}{cdec_stations}(.data)
}
\arguments{
\item{.data}{result of a cdec_stations() call}

\item{...}{named arguments passed into leaflet::addCircleMarkers}
}
\description{
Populate a leaflet map with the results of cdec_stations() call. The function
makes use of leaflet, and so will work only if this is installed on the system.
This function is bundled simply for exploration purposes, it is highly suggested
to make use of leaflet for production maps.
}
\examples{
if (interactive()) {
    cdec_stations(county = "alameda") \%>\% map_stations(label=~name, popup=~station_id)
}
}
