% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{map}
\alias{map}
\title{Map search results}
\usage{
map(x, ...)
}
\arguments{
\item{...}{named arguments passed into leaflet::addCircleMarkers}

\item{.data}{result of a cdec_stations() call}
}
\description{
Populate a leaflet map with the results of a cdec_* call. The function
makes use of leaflet, and so will work only if this is installed on the system.
This function is bundled simply for exploration purposes, it is highly suggested
to make use of leaflet for production maps.
}
\examples{
if (interactive()) {
    cdec_stations(county = "alameda") \%>\% map()
}
}
