% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_data.R
\name{parse_avail_resp}
\alias{parse_avail_resp}
\title{Function parses a response from CDEC avaialable data service}
\usage{
parse_avail_resp(.d_raw)
}
\arguments{
\item{.d_raw}{a response table from cdec via the function call_cdec_avail_service}
}
\value{
parsed version of table returned from cdec
}
\description{
Function parses a response from CDEC avaialable data service
}
