% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_metadata.R
\name{parse_table}
\alias{parse_table}
\title{function parses the results obtained from a cdec call}
\usage{
parse_table(.d)
}
\arguments{
\item{.d}{a data frame containing calls to cdec metadata as rows}
}
\description{
function parses the results obtained from a cdec call
}
