% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{query_metadata}
\alias{query_metadata}
\title{Use three letter code to retieve associated attributes}
\usage{
query_metadata(stations)
}
\arguments{
\item{stations}{a vector of three letter station codes.}
}
\value{
a dataframe with rows as stations submitted to the query
}
\description{
Use three letter code to retieve associated attributes
}
