% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_params.R
\name{rename_params}
\alias{rename_params}
\title{Function takes in a dataframe retrieved from CDECRetrieve::retrieve_station_data
and renames columns to be more readable}
\usage{
rename_params(.d)
}
\arguments{
\item{.d}{dataframe obtained from calling CDECRetrieve::retrieve_station_data}
}
\description{
Function takes in a dataframe retrieved from CDECRetrieve::retrieve_station_data
and renames columns to be more readable
}
