% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sac_valley_wy_types}
\alias{sac_valley_wy_types}
\title{Sacramento Valley Water Year Index}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 111 rows and 6 columns.}
\usage{
sac_valley_wy_types
}
\description{
Chronological Reconstructed Sacramento Valley Water Year Hydrologic Classification Indices
Based on measured unimpaired runoff (in million acre-feet), subject to revision.
}
\examples{
sac_valley_wy_types

}
\keyword{datasets}
