% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{san_joaquin_wy_types}
\alias{san_joaquin_wy_types}
\title{San Joaquin Valley Water Year Index}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 111 rows and 6 columns.}
\usage{
san_joaquin_wy_types
}
\description{
Chronological Reconstructed San Joaquin Valley Water Year Hydrologic Classification Indices
Based on measured unimpaired runoff (in million acre-feet), subject to revision.
}
\keyword{datasets}
