% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{search_sensors}
\alias{search_sensors}
\title{Function does a simple search for terms in description}
\usage{
search_sensors(term)
}
\arguments{
\item{term}{a search term to query for}
}
\value{
data frame with filtered search results
}
\description{
Function does a simple search for terms in description
}
\examples{
temperature_sensors <- search_sensors("temperature")

}
