% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sensors}
\alias{sensors}
\title{The set of CDEC Codes as reported by CDEC}
\format{A data frame cotnaining description to sensors used in CDEC
\describe{
    \item{sensor_no}{sensor number used in CDEC gage data}
    \item{sensor_name}{sensor name corresponding to sensor number}
    \item{description}{short description of the sensor}
    \item{unit}{the unit of measurement used for the sensor}
}}
\source{
\url{http://cdec.com}
}
\description{
A dataset containing all of the cdec codes, along with
other useful iformation including the units of measure for each
}
\keyword{data}
