% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consts.R
\docType{data}
\name{shef_code_lookup}
\alias{shef_code_lookup}
\title{Look up table to go from SHEF Code ---> CDEC Param Code}
\format{An object of class \code{list} of length 37.}
\usage{
shef_code_lookup
}
\description{
Look up table to go from SHEF Code ---> CDEC Param Code
}
\keyword{datasets}
