% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_station_data.R
\name{shef_to_tidy}
\alias{shef_to_tidy}
\title{Function converts shef downloaded data into one of a tidy format}
\usage{
shef_to_tidy(file)
}
\arguments{
\item{file}{shef filename obtained from CDEC}
}
\value{
data frame in tidy form
}
\description{
Function converts shef downloaded data into one of a tidy format
}
