% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{CDEC Stations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2671 rows and 2 columns.}
\usage{
stations
}
\description{
List of all stations reporting data through CDEC. Last updated 2018-06-08
}
\keyword{datasets}
