% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{cdec_available}
\alias{cdec_available}
\title{Function returns a dataframe of the available data}
\usage{
cdec_available(station)
}
\arguments{
\item{station}{a three letter cdec station if}
}
\value{
dataframe of available data for station
}
\description{
Function returns a dataframe of the available data
}
