% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{cdec_query}
\alias{cdec_query}
\title{Function queries the CDEC services to obtain desired station data.}
\usage{
cdec_query(stations, sensor_num, dur_code, start_date, end_date = "")
}
\arguments{
\item{stations}{three letter identification for CDEC location (example "KWK", "SAC", "CCR")}

\item{sensor_num}{sensor number for the measure of interest. (example "20", "01", "25")}

\item{dur_code}{duration code for measure interval, "E", "H", "D", which correspong to Event, Hourly and Daily.}

\item{start_date}{date to start the query on.}

\item{end_date}{a date to end query on, defaults to current date.}
}
\value{
tidy dataframe
}
\description{
User must know beforehand if a given location uses the desired
sensor number as well as duration value. Use show_avaialable_data() to view
an update list of a stations data.
}
\examples{
kwk_hourly_flows <- CDECRetrieve::retrieve_station_data("KWK", "20", "H", "2017-01-01")

}
