% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating-tables.R
\name{cdec_rt}
\alias{cdec_rt}
\title{Get a rating table
Use  station id to find the rating table for the station}
\usage{
cdec_rt(station_id)
}
\arguments{
\item{station_id}{three letter CDEC station id}
}
\value{
dataframe of rating table
}
\description{
Get a rating table
Use  station id to find the rating table for the station
}
