% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating-tables.R
\name{cdec_rt_list}
\alias{cdec_rt_list}
\title{List Rating Tables}
\usage{
cdec_rt_list(station_id = NULL)
}
\arguments{
\item{station_id}{station for the location to get rating description for.}
}
\description{
Get a list of all rating tables available through CDEC
}
\examples{
# list all rating tables in CDEC, you can use filter to search
\dontrun{
cdec_rt_list()
}
}
