% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{cdec_stations}
\alias{cdec_stations}
\title{search for stations in the cdec system}
\usage{
cdec_stations(station_id = NULL, nearby_city = NULL, river_basin = NULL,
  hydro_area = NULL, county = NULL)
}
\arguments{
\item{station_id}{string three letter station code}

\item{nearby_city}{string search stations near supplied city}

\item{river_basin}{string search stations in supplied basin}

\item{hydro_area}{string search stations in supplied hydrological area}

\item{county}{string search stations in supplied county}
}
\description{
search the stations in thec CDEC system using the CDEC Station Search
service \href{https://cdec.water.ca.gov/cgi-progs/staSearch}{here}
}
