% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{map_stations}
\alias{map_stations}
\title{Map Station Search}
\usage{
map_stations(.data, ...)
}
\arguments{
\item{.data}{result of a cdec_stations() call}
}
\value{
map in viewer
}
\description{
Populate a leaflet map with the results of cdec_stations() call
}
