% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_cdec_retrieve.R
\name{retrieve_station_data}
\alias{retrieve_station_data}
\title{Function queries the CDEC services to obtain desired station data}
\usage{
retrieve_station_data(station_id, sensor_num, dur_code, start_date,
  end_date = Sys.Date(), base_url = "shef")
}
\arguments{
\item{station_id}{three letter identification for CDEC location}

\item{sensor_num}{sensor number for the measure of interest}

\item{dur_code}{duration code for measure interval, "E", "H", "D"}

\item{start_date}{date to start the query on}

\item{end_date}{a non-inclusive date to end the query on}
}
\value{
tidy dataframe
}
\description{
Function queries the CDEC services to obtain desired station data
}
