% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_data.R
\name{show_available_data}
\alias{show_available_data}
\title{Function returns a dataframe of the available data}
\usage{
show_available_data(station_id)
}
\description{
Function returns a dataframe of the available data
}
